requirements.txt
ansible==8.7.0
ansible-compat==4.1.11
ansible-core==2.15.9
ansible-lint==6.22.2
ansible-runner==2.3.5
archspec @ file:///croot/archspec_1697725767277/work
attrs==23.2.0
black==24.1.1
boltons @ file:///croot/boltons_1677628692245/work
bracex==2.4
Brotli @ file:///tmp/abs_ecyw11_7ze/croots/recipe/brotli-split_1659616059936/work
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1700303426725/work/certifi
cffi @ file:///croot/cffi_1700254295673/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
Click==7.0
conda @ file:///home/conda/feedstock_root/build_artifacts/conda_1701731653463/work
conda-content-trust @ file:///croot/conda-content-trust_1693490622020/work
conda-libmamba-solver @ file:///croot/conda-libmamba-solver_1702997573971/work/src
conda-package-handling @ file:///croot/conda-package-handling_1690999929514/work
conda_package_streaming @ file:///croot/conda-package-streaming_1690987966409/work
cryptography==41.0.4
distro @ file:///croot/distro_1701455004953/work
docutils==0.20.1
filelock==3.13.1
idna @ file:///croot/idna_1666125576474/work
importlib-metadata==6.2.1
importlib-resources==5.0.7
Jinja2==3.1.2
jmespath==1.0.1
jsonpatch @ file:///tmp/build/80754af9/jsonpatch_1615747632069/work
jsonpointer==2.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
libmambapy @ file:///croot/mamba-split_1698782620632/work/libmambapy
lockfile==0.12.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
menuinst @ file:///croot/menuinst_1702390294373/work
mypy-extensions==1.0.0
netaddr==0.9.0
nujnus @ file:///home/nujnus/workspace/workspace_for_devops/working/nujnus
packaging @ file:///croot/packaging_1693575174725/work
passlib==1.7.4
pathspec==0.12.1
pbr==5.11.1
pexpect==4.9.0
platformdirs @ file:///croot/platformdirs_1692205439124/work
pluggy @ file:///tmp/build/80754af9/pluggy_1648024445381/work
ptyprocess==0.7.0
pycosat @ file:///croot/pycosat_1696536503704/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
Pygments==2.17.2
pyOpenSSL @ file:///croot/pyopenssl_1690223430423/work
PySocks @ file:///tmp/build/80754af9/pysocks_1605305812635/work
python-daemon==3.0.1
PyYAML @ file:///home/conda/feedstock_root/build_artifacts/pyyaml_1695373447169/work
referencing==0.33.0
requests @ file:///croot/requests_1690400202158/work
resolvelib @ file:///home/conda/feedstock_root/build_artifacts/resolvelib_1678667915605/work
rich==13.7.0
rpds-py==0.17.1
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
six==1.16.0
subprocess-tee==0.4.1
tomli==2.0.1
tqdm @ file:///croot/tqdm_1679561862951/work
typing_extensions==4.9.0
urllib3 @ file:///croot/urllib3_1698257533958/work
wcmatch==8.5
yamllint==1.34.0
zipp==3.17.0
zstandard @ file:///croot/zstandard_1677013143055/work
